Start building this package for macOS
When `freeglut` was initially added to conda-forge, there were not X11
packages to rely on and Apple had dropped XQuartz from the OS. So there
was not a really good story for X11 on macOS, which is necessary for
building and using `freeglut` on UNIX generally. However that has since
changed and now we have many of the X11 packages we need already in this
recipe. Hence we start building this package for macOS.